[softmax-xent] Re-organize and enhance test cases
