--- test/test_patch_segment_reduction.py
@@ -1,828 +1,0 @@
-# Copyright 2020 NVIDIA Corporation. All Rights Reserved
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ========================================================================
-
-# Copyright 2015 The TensorFlow Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-"""Functional tests for segment reduction ops."""
-
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
-import itertools
-import os
-import sys
-import unittest
-
-import numpy as np
-import tensorflow as tf
-from tensorflow.python.eager import context
-from tensorflow.python.eager import backprop
-from tensorflow.python.client import session
-from tensorflow.python.framework import constant_op
-from tensorflow.python.framework import dtypes as dtypes_lib
-from tensorflow.python.framework import ops
-from tensorflow.python.framework import test_util
-from tensorflow.python.ops import array_ops
-from tensorflow.python.ops import gradient_checker
-from tensorflow.python.ops import gradient_checker_v2
-from tensorflow.python.ops import gradients_impl
-from tensorflow.python.ops import math_ops
-from tensorflow.python.ops import variables
-from tensorflow.python.platform import test
-
-sys.path.insert(0, '..')
-import fwd9m.tensorflow as fwd9m_tensorflow
-import utils
-
-os.environ['TF_CPP_MIN_LOG_LEVEL'] = '2' # Simplifies logging
-
-# Notes:
-# 1. The ops were expected to operate deterministically on the CPU and they do
-#    indeed operate deterministically if forcely pinned to the CPU with
-#    tf.device('/device:CPU:0'). What is not fully understood is why when they
-#    are placed on the CPU using self.session(use_gpu=False), the ops still
-#    introduce nondeterminism. By setting the log_device_placement parameter in
-#    the session config to True under these conditions, we are able to confirm
-#    that the ops are running on the CPU.
-# 2. To capture nondeterminism, random input data is necessary.
-# 3. The nondeterminism of dtypes_lib.float64, dtypes_lib.complex128 cannot be
-#    removed by this patch, so they are not tested.
-# 4. The regular op tests below, represented by all the test classes except the
-#    final two, which have names ending in "Deterministic", were taken from
-#    tensorflow/python/kernel_tests/segment_reduction_ops_test.py
-#    (as of 2020-08-02); URL to file-at-commit:
-#    https://github.com/tensorflow/tensorflow/blob/6371d4a38cfb122a8d9b2a03d5f56444e95462b0/tensorflow/python/kernel_tests/segment_reduction_ops_test.py
-# 5. The names of most of the upstream test classes are confusing (even more so
-#    in the context of their limited use here), so the names have been changed
-#    here, as appropriate, along with comments to indicate the original test
-#    class names.
-
-
-class SegmentReductionHelper(test.TestCase):
-
-  def _random_input(self, input_shape, dtype=dtypes_lib.int32):
-    np.random.seed(hash(dtype) % 256)
-
-    np_values = np.random.random(input_shape).astype(dtype.as_numpy_dtype)
-    # Add a non-zero imaginary component to complex types.
-    if dtype.is_complex:
-      np_values -= 1j * np_values
-    return constant_op.constant(
-        np_values, shape=input_shape, dtype=dtype), np_values
-
-  def _input(self, input_shape, dtype=dtypes_lib.int32):
-    num_elem = 1
-    for x in input_shape:
-      num_elem *= x
-    values = np.arange(1, num_elem + 1)
-    np_values = values.reshape(input_shape).astype(dtype.as_numpy_dtype)
-    # Add a non-zero imaginary component to complex types.
-    if dtype.is_complex:
-      np_values -= 1j * np_values
-    return constant_op.constant(
-        np_values, shape=input_shape, dtype=dtype), np_values
-
-  def _segmentReduce(self, indices, x, op1, op2=None, num_segments=None,
-                     initial_value=0):
-    if not x.size:
-      return np.array([])
-    indices = np.asarray(indices)
-    if num_segments is None:
-      num_segments = indices[-1] + 1
-    output = [None] * num_segments
-    slice_shape = x.shape[indices.ndim:]
-    x_flat = x.reshape((indices.size,) + slice_shape)
-    for i, index in enumerate(indices.ravel()):
-      if (output[index] is not None) and op1 == np.max:
-        for j in range(0, output[index].shape[0]):
-          output[index][j] = op1([output[index][j], x_flat[i][j]])
-      elif output[index] is not None:
-        output[index] = op1(output[index], x_flat[i])
-      else:
-        output[index] = x_flat[i]
-    # zero initialize values that are still uncalcuated.
-    initial_value_slice = np.ones(slice_shape) * initial_value
-    output = [o if o is not None else initial_value_slice for o in output]
-    if op2 is not None:
-      output = [op2(o) for o in output]
-    output = [o.reshape(slice_shape) for o in output]
-    return np.array(output)
-
-  def _mean_cum_op(self, x, y):
-    return (x[0] + y, x[1] + 1) if isinstance(x, tuple) else (x + y, 2)
-
-  def _mean_reduce_op(self, x):
-    return x[0] / x[1] if isinstance(x, tuple) else x
-
-  def _sqrt_n_reduce_op(self, x):
-    return x[0] / np.sqrt(x[1]) if isinstance(x, tuple) else x
-
-
-# Upstream class name: SegmentReductionOpTest
-class SegmentSumTest(SegmentReductionHelper):
-
-  def testValues(self):
-    dtypes = [
-        dtypes_lib.float32, dtypes_lib.float64, dtypes_lib.int64,
-        dtypes_lib.int32, dtypes_lib.complex64, dtypes_lib.complex128
-    ]
-
-    # Each item is np_op1, np_op2, tf_op
-    ops_list = [(np.add, None, math_ops.segment_sum)]
-
-    # A subset of ops has been enabled for complex numbers
-    complex_ops_list = [(np.add, None, math_ops.segment_sum)]
-
-    n = 10
-    shape = [n, 2]
-    indices = [i // 3 for i in range(n)]
-    for dtype in dtypes:
-      if dtype in (dtypes_lib.complex64, dtypes_lib.complex128):
-        curr_ops_list = complex_ops_list
-      else:
-        curr_ops_list = ops_list
-      for use_gpu in [True, False]:
-        with self.cached_session(use_gpu=use_gpu):
-          tf_x, np_x = self._input(shape, dtype=dtype)
-          for np_op1, np_op2, tf_op in curr_ops_list:
-            np_ans = self._segmentReduce(indices, np_x, np_op1, np_op2)
-            s = tf_op(data=tf_x, segment_ids=indices)
-            tf_ans = self.evaluate(s)
-            self.assertAllClose(np_ans, tf_ans)
-            # NOTE(mrry): The static shape inference that computes
-            # `tf_ans.shape` can only infer that sizes from dimension 1
-            # onwards, because the size of dimension 0 is data-dependent
-            # and may therefore vary dynamically.
-            self.assertAllEqual(np_ans.shape[1:], tf_ans.shape[1:])
-
-  @test_util.run_deprecated_v1
-  def testSegmentIdsShape(self):
-    shape = [4, 4]
-    tf_x, _ = self._input(shape)
-    indices = constant_op.constant([0, 1, 2, 2], shape=[2, 2])
-    with self.assertRaises(ValueError):
-      math_ops.segment_sum(data=tf_x, segment_ids=indices)
-
-  @test_util.run_deprecated_v1
-  def testSegmentIdsSize(self):
-    shape = [4, 4]
-    for use_gpu in [True, False]:
-      with self.cached_session(use_gpu=use_gpu):
-        tf_x, _ = self._input(shape)
-        indices = [0, 1]
-        s = math_ops.segment_sum(data=tf_x, segment_ids=indices)
-        with self.assertRaisesOpError("segment_ids should be the same size"):
-          self.evaluate(s)
-
-  @test_util.run_deprecated_v1
-  def testSegmentIdsValid(self):
-    # This is a baseline for the following SegmentIdsInvalid* tests.
-    shape = [4, 4]
-    for use_gpu in [True, False]:
-      with self.cached_session(use_gpu=use_gpu):
-        tf_x, _ = self._input(shape, dtype=dtypes_lib.float32)
-        indices = [0, 0, 0, 1]
-        result = math_ops.segment_sum(data=tf_x, segment_ids=indices).eval()
-        self.assertAllEqual([[15, 18, 21, 24], [13, 14, 15, 16]], result)
-
-  def testSegmentIdsGreaterThanZero(self):
-    shape = [4, 4]
-    for use_gpu in [True, False]:
-      with self.cached_session(use_gpu=use_gpu):
-        tf_x, np_x = self._input(shape, dtype=dtypes_lib.float32)
-        indices = [1, 1, 2, 2]
-        np_ans = self._segmentReduce(indices, np_x, np.add)
-        s = math_ops.segment_sum(data=tf_x, segment_ids=indices)
-        tf_ans = self.evaluate(s)
-        self.assertAllClose(np_ans, tf_ans)
-
-  def testSegmentIdsHole(self):
-    shape = [4, 4]
-    for use_gpu in [True, False]:
-      with self.cached_session(use_gpu=use_gpu):
-        tf_x, np_x = self._input(shape, dtype=dtypes_lib.float32)
-        indices = [0, 0, 3, 3]
-        np_ans = self._segmentReduce(indices, np_x, np.add)
-        s = math_ops.segment_sum(data=tf_x, segment_ids=indices)
-        tf_ans = self.evaluate(s)
-        self.assertAllClose(np_ans, tf_ans)
-
-  @test_util.run_deprecated_v1
-  def testSegmentIdsInvalid1(self):
-    shape = [4, 4]
-    with self.cached_session():
-      tf_x, _ = self._input(shape)
-      indices = [-1, -1, 0, 0]
-      s = math_ops.segment_sum(data=tf_x, segment_ids=indices)
-      with self.assertRaisesOpError(
-          r"Segment id -1 out of range \[0, 1\), possibly because "
-          "'segment_ids' input is not sorted."):
-        self.evaluate(s)
-
-  @test_util.run_deprecated_v1
-  def testSegmentIdsInvalid2(self):
-    shape = [4, 4]
-    with self.cached_session():
-      tf_x, _ = self._input(shape)
-      indices = [0, 1, 0, 1]
-      s = math_ops.segment_sum(data=tf_x, segment_ids=indices)
-      with self.assertRaisesOpError("segment ids are not increasing"):
-        self.evaluate(s)
-
-  @test_util.run_deprecated_v1
-  def testSegmentIdsInvalid3(self):
-    shape = [4, 4]
-    with self.cached_session():
-      tf_x, _ = self._input(shape)
-      indices = [0, 1, 2, 0]
-      s = math_ops.segment_sum(data=tf_x, segment_ids=indices)
-      with self.assertRaisesOpError(
-          r"Segment id 1 out of range \[0, 1\), possibly "
-          "because 'segment_ids' input is not sorted."):
-        self.evaluate(s)
-
-  @test_util.run_deprecated_v1
-  def testSegmentIdsInvalid4(self):
-    shape = [4, 4]
-    for use_gpu in [True, False]:
-      with self.cached_session(use_gpu=use_gpu):
-        tf_x, _ = self._input(shape, dtype=dtypes_lib.float32)
-        indices = [0, 0, 0, -1]
-        s = math_ops.segment_sum(data=tf_x, segment_ids=indices)
-        with self.assertRaisesOpError("segment ids must be >= 0"):
-          self.evaluate(s)
-
-  @test_util.run_deprecated_v1
-  def testSegmentIdsInvalid5(self):
-    shape = [4, 4]
-    for use_gpu in [True, False]:
-      with self.cached_session(use_gpu=use_gpu):
-        tf_x, _ = self._input(shape, dtype=dtypes_lib.float32)
-        indices = [0, 0, 0, -2]
-        s = math_ops.segment_sum(data=tf_x, segment_ids=indices)
-        with self.assertRaisesOpError("segment ids must be >= 0"):
-          self.evaluate(s)
-
-  @test_util.run_deprecated_v1
-  def testGradient(self):
-    shape = [4, 4]
-    indices = [0, 1, 2, 2]
-    for tf_op in [
-        math_ops.segment_sum]:
-      with self.cached_session():
-        tf_x, np_x = self._input(shape, dtype=dtypes_lib.float64)
-        s = tf_op(data=tf_x, segment_ids=indices)
-        jacob_t, jacob_n = gradient_checker.compute_gradient(
-            tf_x,
-            shape,
-            s, [3, 4],
-            x_init_value=np_x.astype(np.double),
-            delta=1)
-      self.assertAllClose(jacob_t, jacob_n)
-
-  # Method removed because it only tests math_ops.segment_mean
-  # def testDataInvalid(self):
-  #   ...
-
-
-# Upstream class name: UnsortedSegmentTest
-class UnsortedSegmentSumTest(SegmentReductionHelper):
-
-  def __init__(self, methodName='runTest'):
-    # Each item is np_op1, np_op2, tf_op, initial_value functor
-    self.ops_list = [(np.add, None,
-                      math_ops.unsorted_segment_sum, lambda t: 0)]
-
-    # A subset of ops has been enabled for complex numbers
-    self.complex_ops_list = [(np.add, None,
-                              math_ops.unsorted_segment_sum, lambda t: 0)]
-    self.differentiable_dtypes = [dtypes_lib.float16, dtypes_lib.float32,
-                                  dtypes_lib.float64]
-    self.all_dtypes = (self.differentiable_dtypes +
-                       [dtypes_lib.bfloat16,
-                        dtypes_lib.int64, dtypes_lib.int32,
-                        dtypes_lib.complex64, dtypes_lib.complex128])
-    super(UnsortedSegmentSumTest, self).__init__(methodName=methodName)
-
-  def testValues(self):
-    indices_flat = np.array([0, 4, 0, 8, 3, 8, 4, 7, 7, 3])
-    num_segments = 12
-    for indices in indices_flat, indices_flat.reshape(5, 2):
-      shape = indices.shape + (2,)
-      for dtype in self.all_dtypes:
-        ops_list = self.complex_ops_list if dtype.is_complex else self.ops_list
-        tf_x, np_x = self._input(shape, dtype=dtype)
-        for use_gpu in [True, False]:
-          with self.cached_session(use_gpu=True):
-            for np_op1, np_op2, tf_op, init_op in ops_list:
-              # sqrt_n doesn't support integers
-              if (np_op2 == self._sqrt_n_reduce_op and dtype.is_integer):
-                continue
-              # todo(philjd): enable this test once real_div supports bfloat16
-              if (np_op2 in [self._sqrt_n_reduce_op, self._mean_reduce_op] and
-                  dtype == dtypes_lib.bfloat16):
-                continue
-              np_ans = self._segmentReduce(
-                  indices, np_x, np_op1, np_op2, num_segments=num_segments,
-                  initial_value=init_op(dtype))
-              s = tf_op(tf_x, segment_ids=indices, num_segments=num_segments)
-              tf_ans = self.evaluate(s)
-              if dtype is dtypes_lib.bfloat16:
-                tf_ans = tf_ans.astype(np.float32)
-              self.assertAllCloseAccordingToType(np_ans, tf_ans)
-              self.assertShapeEqual(np_ans, s)
-
-  def testNumSegmentsTypes(self):
-    dtypes = [dtypes_lib.int32, dtypes_lib.int64]
-    indices_flat = np.array([0, 4, 0, 8, 3, 8, 4, 7, 7, 3])
-    num_segments = 12
-    for indices in indices_flat, indices_flat.reshape(5, 2):
-      shape = indices.shape + (2,)
-      for dtype in dtypes:
-        with self.cached_session(use_gpu=True):
-          tf_x, np_x = self._input(shape)
-          num_segments_constant = constant_op.constant(
-              num_segments, dtype=dtype)
-          np_ans = self._segmentReduce(
-              indices, np_x, np.add, op2=None, num_segments=num_segments)
-          s = math_ops.unsorted_segment_sum(
-              data=tf_x,
-              segment_ids=indices,
-              num_segments=num_segments_constant)
-          tf_ans = self.evaluate(s)
-        self.assertAllClose(np_ans, tf_ans)
-        self.assertShapeEqual(np_ans, s)
-
-  @test_util.run_deprecated_v1
-  def testGradientsTFGradients(self):
-    num_cols = 2
-    indices_flat = np.array([0, 4, 0, -1, 3, -1, 4, 7, 7, 3])
-    num_segments = max(indices_flat) + 3
-    for dtype in self.differentiable_dtypes:
-      ops_list = self.complex_ops_list if dtype.is_complex else self.ops_list
-      for indices in indices_flat, indices_flat.reshape(5, 2):
-        shape = indices.shape + (num_cols,)
-        # test CPU and GPU as tf.gather behaves differently on each device
-        for use_gpu in [False, True]:
-          with self.cached_session(use_gpu=use_gpu):
-            for _, _, tf_op, _ in ops_list:
-              tf_x, np_x = self._input(shape, dtype=dtype)
-              s = tf_op(tf_x, indices, num_segments)
-              jacob_t, jacob_n = gradient_checker.compute_gradient(
-                  tf_x,
-                  shape,
-                  s, [num_segments, num_cols],
-                  x_init_value=np_x,
-                  delta=1.)
-              self.assertAllCloseAccordingToType(jacob_t, jacob_n,
-                                                 half_atol=1e-2)
-
-  def _computeGradient(self, tf_op, indices, num_segments,
-                       shape, num_cols, dtype):
-    tf_x, np_x = self._input(shape, dtype=dtype)
-    if context.executing_eagerly():
-      def f(x):
-        return tf_op(x, indices, num_segments)
-
-      gradient_tape_jacob_t, jacob_n = gradient_checker_v2.compute_gradient(
-          f, [tf_x], delta=1.0)
-      self.assertAllClose(jacob_n, gradient_tape_jacob_t)
-    else:
-       with self.cached_session():
-         s = tf_op(tf_x, indices, num_segments)
-         jacob_t, jacob_n = gradient_checker.compute_gradient(
-            tf_x,
-            shape,
-            s, [num_segments, num_cols],
-            x_init_value=np_x,
-            delta=1)
-         self.assertAllClose(jacob_t, jacob_n)
-
-  # This method has been enhanced to run on older versions of TensorFlow
-  @test_util.run_in_graph_and_eager_modes
-  def testGradientsGradientTape(self):
-    num_cols = 2
-    indices_flat = np.array([0, 4, 0, -1, 3, -1, 4, 7, 7, 3])
-    num_segments = max(indices_flat) + 3
-    for dtype in self.differentiable_dtypes:
-      ops_list = self.complex_ops_list if dtype.is_complex else self.ops_list
-      for indices in indices_flat, indices_flat.reshape(5, 2):
-        shape = indices.shape + (num_cols,)
-        # test CPU and GPU as tf.gather behaves differently on each device
-        for use_gpu in [test_util.use_gpu, test_util.force_cpu]:
-          with use_gpu():
-            for _, _, tf_op, _ in ops_list:
-              self._computeGradient(tf_op, indices, num_segments, shape,
-                                    num_cols, dtype)
-
-  # Method removed because it only tests math_ops.unsorted_segment_prod
-  # def testProdGrad(self):
-  #   ...
-
-  @test_util.run_deprecated_v1
-  def testGradientMatchesSegmentSum(self):
-    # Strategy: compute the gradient for UnsortedSegmentSum and SegmentSum
-    # and compare the outputs, which should be identical.
-    # NB: for this test to work, indices must be valid for SegmentSum, namely
-    # it must be sorted, the indices must be contiguous, and num_segments
-    # must be max(indices) + 1.
-    indices = [0, 0, 1, 1, 1, 2, 3, 4, 5]
-    n = len(indices)
-    num_cols = 2
-    shape = [n, num_cols]
-    num_segments = max(indices) + 1
-    for dtype in self.differentiable_dtypes:
-      with self.cached_session(use_gpu=True):
-        tf_x, np_x = self._input(shape, dtype=dtype)
-        # Results from UnsortedSegmentSum
-        unsorted_s = math_ops.unsorted_segment_sum(
-            data=tf_x, segment_ids=indices, num_segments=num_segments)
-        unsorted_jacob_t, unsorted_jacob_n = (
-            gradient_checker.compute_gradient(tf_x, shape, unsorted_s,
-                                              [num_segments, num_cols],
-                                              x_init_value=np_x, delta=1))
-
-        # Results from SegmentSum
-        sorted_s = math_ops.segment_sum(data=tf_x, segment_ids=indices)
-        sorted_jacob_t, sorted_jacob_n = gradient_checker.compute_gradient(
-            tf_x,
-            shape,
-            sorted_s, [num_segments, num_cols],
-            x_init_value=np_x,
-            delta=1)
-      self.assertAllClose(unsorted_jacob_t, sorted_jacob_t)
-      self.assertAllClose(unsorted_jacob_n, sorted_jacob_n)
-
-  @test_util.run_deprecated_v1
-  def testBadIndices(self):
-    # Note: GPU kernel does not return the out-of-range error needed for this
-    # test, so this test is marked as cpu-only.
-    # Note: With PR #13055 a negative index will be ignored silently.
-    with self.session(use_gpu=False):
-      for bad in [[2]], [[7]]:
-        unsorted = math_ops.unsorted_segment_sum([[17]], bad, num_segments=2)
-        with self.assertRaisesOpError(
-            r"segment_ids\[0,0\] = %d is out of range \[0, 2\)" % bad[0][0]):
-          self.evaluate(unsorted)
-
-  @test_util.run_deprecated_v1
-  def testEmptySecondDimension(self):
-    dtypes = [np.float16, np.float32, np.float64, np.int64, np.int32,
-              np.complex64, np.complex128]
-    with self.session(use_gpu=True):
-      for dtype in dtypes:
-        for itype in (np.int32, np.int64):
-          data = np.zeros((2, 0), dtype=dtype)
-          segment_ids = np.array([0, 1], dtype=itype)
-          unsorted = math_ops.unsorted_segment_sum(data, segment_ids, 2)
-          self.assertAllEqual(unsorted.eval(), np.zeros((2, 0), dtype=dtype))
-
-  def testDropNegatives(self):
-    # Note: the test is done by replacing segment_ids with 8 to -1
-    # for index  and replace values generated by numpy with 0.
-    indices_flat = np.array([0, 4, 0, 8, 3, 8, 4, 7, 7, 3])
-    num_segments = 12
-    for indices in indices_flat, indices_flat.reshape(5, 2):
-      shape = indices.shape + (2,)
-      for dtype in self.all_dtypes:
-        with self.session(use_gpu=True):
-          tf_x, np_x = self._input(shape, dtype=dtype)
-          np_ans = self._segmentReduce(
-              indices, np_x, np.add, op2=None, num_segments=num_segments)
-          # Replace np_ans[8] with 0 for the value
-          np_ans[8:] = 0
-          # Replace 8 with -1 in indices
-          np.place(indices, indices == 8, [-1])
-          s = math_ops.unsorted_segment_sum(
-              data=tf_x, segment_ids=indices, num_segments=num_segments)
-          tf_ans = self.evaluate(s)
-        self.assertAllClose(np_ans, tf_ans)
-        self.assertShapeEqual(np_ans, s)
-
-
-class SegmentReductionOpBenchmark(test.Benchmark):
-
-  outer_dim_options = [2**x for x in range(9, 14, 2)]
-  ratio_options = [2**x for x in range(1, 6, 2)]
-  inner_dim_options = [2**x for x in range(9, 14, 2)]
-  # randomly generated sizes with less alignments
-  inner_dim_options += [
-      1120, 1215, 1856, 1302, 1329, 1531, 1313, 1672, 1851, 1584
-  ]
-  dtype_options = [np.float32, np.float64]
-  options = (outer_dim_options, ratio_options, inner_dim_options, dtype_options)
-  # pylint: disable=g-long-lambda
-  op_functors = [lambda vc, vs, seg_ids:
-                 ("sorted", math_ops.segment_sum(vc, vs)),
-                 lambda vc, vs, seg_ids:
-                 ("unsorted",
-                  math_ops.unsorted_segment_sum(vc, vs, seg_ids[-1]+1))]
-  # pylint: enable=g-long-lambda
-  repeat = 10
-
-  def _npTypeToStr(self, t):
-    if t == np.float32:
-      return "fp32"
-    if t == np.float64:
-      return "fp64"
-
-  def _runGraph(self, op_functor, outer_dim, ratio, inner_dim, dtype):
-    output_outer_dim = int(outer_dim / ratio)
-    const = np.random.randint(5, size=(outer_dim, inner_dim))
-    seg_ids = np.sort(np.random.randint(output_outer_dim, size=outer_dim))
-    vs = variables.Variable(seg_ids.astype(np.int32))
-    with ops.device("/gpu:0"):
-      vc = variables.Variable(const.astype(dtype))
-    name, op = op_functor(vc, vs, seg_ids)
-    with session.Session() as sess:
-      variables.global_variables_initializer().run()
-      r = self.run_op_benchmark(
-          sess,
-          op,
-          min_iters=self.repeat,
-          name="_".join(
-              map(str,
-                  [name, outer_dim, ratio, inner_dim,
-                   self._npTypeToStr(dtype)])))
-    return name, r["wall_time"]
-
-  def benchmarkSegmentSumGPU(self):
-    if not test.is_gpu_available(cuda_only=True):
-      return
-    for outer_dim, ratio, inner_dim, dtype in itertools.product(*self.options):
-      op_functor = self.op_functors[0]
-      with ops.Graph().as_default():
-        self._runGraph(op_functor, outer_dim, ratio, inner_dim, dtype)
-
-  def benchmarkUnsortedSegmentSumGPU(self):
-    if not test.is_gpu_available(cuda_only=True):
-      return
-    for outer_dim, ratio, inner_dim, dtype in itertools.product(*self.options):
-      op_functor = self.op_functors[1]
-      with ops.Graph().as_default():
-        self._runGraph(op_functor, outer_dim, ratio, inner_dim, dtype)
-
-
-class SegmentSumDeterministicTest(SegmentReductionHelper):
-
-  def __init__(self, methodName='runTest'):
-  # Each item is np_op1, np_op2, tf_op, initial_value functor
-    self.ops_list = [(np.add, None,
-                      math_ops.segment_sum, lambda t: 0),
-                     (np.add, None,
-                      tf.math.segment_sum, lambda t: 0)]
-
-    # A subset of ops has been enabled for complex numbers
-    self.complex_ops_list = [(np.add, None,
-                              math_ops.segment_sum, lambda t: 0),
-                             (np.add, None,
-                              tf.math.segment_sum, lambda t: 0)]
-
-    self.differentiable_dtypes = [dtypes_lib.float16, dtypes_lib.float32]
-
-    self.all_dtypes = (self.differentiable_dtypes +
-                      [dtypes_lib.bfloat16,
-                        dtypes_lib.int64, dtypes_lib.int32,
-                        dtypes_lib.complex64])
-    self.repeat_count = 5
-    super(SegmentSumDeterministicTest,
-          self).__init__(methodName=methodName)
-
-  def _testForwardCase(self, dtype, indices, ops_list, shape):
-    # have to use float to exec nond9m
-    tf_x, _ = self._random_input(shape, dtype=dtype)
-    # with utils.force_gpu_session(self):
-    with self.session(use_gpu=True):
-      for _, _, tf_op, _ in ops_list:
-        run_ref = tf_op(data=tf_x, segment_ids=indices, name="tf_op_output")
-        for i in range(self.repeat_count):
-          self.assertAllEqual(tf_op(data=tf_x, segment_ids=indices), run_ref)
-
-  def _testBackwardCase(self, dtype, indices, tf_op, shape):
-    numpy_seed = 123
-
-    def _randomDataOp(shape, data_type, seed):
-      if seed is not None:
-        np.random.seed(seed)
-      return constant_op.constant(np.random.random_sample(shape),
-                                  dtype=data_type)
-
-    input_val = _randomDataOp(shape, dtype, seed=None)
-    output_shape = [indices[-1]+1, shape[1]]
-    if context.executing_eagerly():
-      def op_gradients(local_seed):
-        with backprop.GradientTape() as tape:
-          tape.watch(input_val)
-          op_output = tf_op(input_val, indices)
-          upstream_gradients = _randomDataOp(output_shape, dtype, local_seed)
-          gradient_injector_output = op_output * upstream_gradients
-        return tape.gradient(gradient_injector_output, input_val)
-
-      for i in range(self.repeat_count):
-        local_seed = numpy_seed + i # select different upstream gradients
-        result_a = op_gradients(local_seed)
-        result_b = op_gradients(local_seed)
-        self.assertAllEqual(result_a, result_b)
-
-    else:
-      op_output = tf_op(input_val, indices)
-      upstream_gradients = array_ops.placeholder(dtype, shape=output_shape,
-                                                 name='upstream_gradients')
-      gradient_injector_output = op_output * upstream_gradients
-      op_gradients = gradients_impl.gradients(
-            gradient_injector_output,
-            input_val,
-            grad_ys=None,
-            colocate_gradients_with_ops=True)[0]
-
-      for i in range(self.repeat_count):
-        feed_dict = {upstream_gradients:np.random.random(output_shape)}
-        result_a = op_gradients.eval(feed_dict=feed_dict)
-        result_b = op_gradients.eval(feed_dict=feed_dict)
-        self.assertAllEqual(result_a, result_b)
-
-  @test_util.run_in_graph_and_eager_modes
-  def testForward(self):
-    num_cols = 8
-    num_segments = 32
-    segment_size = 256
-
-    shape = [segment_size, num_cols]
-    indices = np.random.randint(low=0, high=num_segments, size=(segment_size,))
-    indices = np.sort(indices)
-
-    for dtype in self.all_dtypes:
-      ops_list = self.complex_ops_list if dtype.is_complex else self.ops_list
-      self._testForwardCase(dtype, indices, ops_list, shape)
-
-  # The backward operation is not known or expected to introduce nondeterminism
-  # but we're testing it for completeness.
-  @test_util.run_in_graph_and_eager_modes
-  def testBackward(self):
-    gradient_test = True
-    num_cols = 8
-    num_segments = 32
-    segment_size = 256
-    shape = [segment_size, num_cols]
-    indices = np.random.randint(low=0, high=num_segments, size=(segment_size,))
-    indices = np.sort(indices)
-
-    with utils.force_gpu_session(self):
-    # with self.session(force_gpu=True):#force_gpu=True leads to XLA issue
-      for dtype in self.differentiable_dtypes:
-        ops_list = self.complex_ops_list if dtype.is_complex else self.ops_list
-        for _, _, tf_op, _ in ops_list:
-          self._testBackwardCase(dtype, indices, tf_op, shape)
-
-
-class UnsortedSegmentSumDeterministicTest(SegmentReductionHelper):
-
-  def __init__(self, methodName='runTest'):
-    # Each item is np_op1, np_op2, tf_op, initial_value functor
-    self.ops_list = [(np.add, None,
-                      math_ops.unsorted_segment_sum, lambda t: 0),
-                     (np.add, None,
-                      tf.math.unsorted_segment_sum, lambda t: 0)]
-
-    # A subset of ops has been enabled for complex numbers
-    self.complex_ops_list = [(np.add, None,
-                              math_ops.unsorted_segment_sum, lambda t: 0),
-                             (np.add, None,
-                              tf.math.unsorted_segment_sum, lambda t: 0)]
-
-    self.differentiable_dtypes = [dtypes_lib.float16, dtypes_lib.float32]
-    self.all_dtypes = (self.differentiable_dtypes +
-                       [dtypes_lib.bfloat16,
-                        dtypes_lib.int64, dtypes_lib.int32,
-                        dtypes_lib.complex64])
-    self.repeat_count = 5
-    super(
-        UnsortedSegmentSumDeterministicTest, self).__init__(
-            methodName=methodName)
-
-  def _testForwardCase(self, dtype, indices, num_segments, num_cols, ops_list,
-                       shape):
-    x, _  = self._random_input(shape, dtype=dtype)
-    def forward(tf_op):
-      s = tf_op(x, indices, num_segments)
-      tf_ans = self.evaluate(s)
-      return tf_ans
-
-    # with utils.force_gpu_session(self):
-    with self.session(use_gpu=True):
-      for _, _, tf_op, _ in ops_list:
-        run_ref = forward(tf_op)
-        for i in range(self.repeat_count):
-          self.assertAllEqual(forward(tf_op), run_ref)
-
-  def _testBackwardCase(self, dtype, indices, num_segments, op_binding, shape):
-    numpy_seed = 123
-    _, _, tf_op, _ = op_binding
-
-    def _randomDataOp(shape, data_type, seed):
-      if seed is not None:
-        np.random.seed(seed)
-      return constant_op.constant(np.random.random_sample(shape),
-                                  dtype=data_type)
-
-    input_val = _randomDataOp(shape, dtype, seed=None)
-
-    if context.executing_eagerly():
-      def op_gradients(local_seed):
-        with backprop.GradientTape() as tape:
-          tape.watch(input_val)
-          op_output = tf_op(input_val, indices, num_segments)
-          upstream_gradients = _randomDataOp(op_output.shape, dtype, local_seed)
-          gradient_injector_output = op_output * upstream_gradients
-        return tape.gradient(gradient_injector_output, input_val)
-
-      for i in range(self.repeat_count):
-        local_seed = numpy_seed + i # select different upstream gradients
-        result_a = op_gradients(local_seed)
-        result_b = op_gradients(local_seed)
-        self.assertAllEqual(result_a, result_b)
-
-    else:
-      op_output = tf_op(input_val, indices, num_segments)
-      output_shape = op_output.shape
-      upstream_gradients = array_ops.placeholder(dtype, shape=output_shape,
-                                                 name='upstream_gradients')
-      gradient_injector_output = op_output * upstream_gradients
-      op_gradients = gradients_impl.gradients(
-            gradient_injector_output,
-            input_val,
-            grad_ys=None,
-            colocate_gradients_with_ops=True)[0]
-
-      for i in range(self.repeat_count):
-        feed_dict = {upstream_gradients:np.random.random(output_shape)}
-        result_a = op_gradients.eval(feed_dict=feed_dict)
-        result_b = op_gradients.eval(feed_dict=feed_dict)
-        self.assertAllEqual(result_a, result_b)
-
-  @test_util.run_in_graph_and_eager_modes
-  def testForward(self):
-    num_cols = 2
-    num_rows = 64
-    num_segments = 64
-    segment_size = num_cols * num_rows
-    indices_flat = np.random.randint(low=-1, high=num_segments,
-                                     size=(segment_size,))
-
-    for dtype in self.all_dtypes:
-      ops_list = self.complex_ops_list if dtype.is_complex else self.ops_list
-      for indices in indices_flat, indices_flat.reshape(num_rows, num_cols):
-        shape = indices.shape + (num_cols,)
-        self._testForwardCase(
-            dtype, indices, num_segments, num_cols, ops_list, shape)
-
-  # The backward operation is not known or expected to introduce nondeterminism
-  # but we're testing it for completeness.
-  @test_util.run_in_graph_and_eager_modes
-  def testBackward(self):
-    num_cols = 2
-    num_rows = 64
-    num_segments = 64
-    segment_size = num_cols * num_rows
-    indices_flat = np.random.randint(low=-1, high=num_segments,
-                                     size=(segment_size,))
-
-    with utils.force_gpu_session(self):
-    # with self.session(force_gpu=True):
-      for dtype in self.differentiable_dtypes:
-        ops_list = self.complex_ops_list if dtype.is_complex else self.ops_list
-        for op_binding in ops_list:
-          for indices in indices_flat, indices_flat.reshape(num_rows, num_cols):
-            shape = indices.shape + (num_cols,)
-            self._testBackwardCase(
-                dtype, indices, num_segments, op_binding, shape)
-
-
-if __name__ == "__main__":
-  fwd9m_tensorflow.enable_determinism()
-  test.main()
